# api/openapi.yaml
openapi: 3.1.0
info:
  title: sshc Local API
  version: 0.1.0
paths:
...
  /api/v1/passwords:
    get:
      operationId: getPasswordVault
      responses:
        "200":
          description: Whether a vault exists, whether it is unlocked, and which hosts have a password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/initialise:
    post:
      operationId: initialisePasswordVault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PassphraseRequest" }
      responses:
        "200":
          description: Vault created and unlocked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/unlock:
    post:
      operationId: unlockPasswordVault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PassphraseRequest" }
      responses:
        "200":
          description: Vault unlocked for this run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /api/v1/passwords/change:
    post:
      operationId: changeMasterPassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeMasterPasswordRequest" }
      responses:
        "200":
          description: The master password was changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChangeMasterPasswordResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/lock:
    post:
      operationId: lockPasswordVault
      responses:
        "200":
          description: Derived key and every pending askpass token forgotten
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/{alias}/eligibility:
    get:
      operationId: getPasswordEligibility
      # 読み取り専用であり、それは意図的である。ここが報告する内容はすべて、この
      # アプリケーションがすでに解析しているファイルから答えられる。パスワードを
      # 保存してよいかを判断するために接続を開く検査は、ユーザーが求めていない
      # 接続になってしまう。
      parameters:
        - name: alias
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "200":
          description: What stands between this host and a stored password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordEligibility" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/{alias}:
    put:
      operationId: storePassword
      parameters:
        - name: alias
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StorePasswordRequest" }
      responses:
        "200":
          description: Password stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: forgetPassword
      parameters:
        - name: alias
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "200":
          description: Password forgotten
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/credentials:
    get:
      operationId: listCredentials
      responses:
        "200":
          description: Credential names and what uses them, never a value
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialList" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/credentials/{kind}/assign:
    put:
      operationId: assignCredential
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignCredentialRequest" }
      responses:
        "200":
          description: The credential lists after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialList" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/credentials/{kind}/assign/{subject}:
    delete:
      operationId: unassignCredential
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: subject, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The credential lists after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialList" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/credentials/{kind}/{name}:
    put:
      operationId: storeCredential
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoreCredentialRequest" }
      responses:
        "200":
          description: The credential lists after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialList" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteCredential
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The credential lists after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialList" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
        path: { type: string }
        line: { type: integer }
        column: { type: integer }
        diagnostics: { type: array, items: { $ref: "#/components/schemas/Diagnostic" } }
        conflict: { $ref: "#/components/schemas/ConflictReport" }
...
    Credential:
      type: object
      additionalProperties: false
      required: [kind, name, uses]
      properties:
        kind: { type: string }
        name: { type: string }
        # この資格情報を指しているもの。パスワードなら alias、パスフレーズなら
        # 鍵のパスである。値そのものは決して含まない。
        uses: { type: array, items: { type: string } }
    CredentialList:
      type: object
      additionalProperties: false
      required: [credentials]
      properties:
        credentials: { type: array, items: { $ref: "#/components/schemas/Credential" } }
...
    ChangeMasterPasswordRequest:
      type: object
      additionalProperties: false
      required: [current, next]
      properties:
        current: { type: string, minLength: 1, maxLength: 1024 }
        next: { type: string, minLength: 1, maxLength: 1024 }
    ChangeMasterPasswordResult:
      type: object
      additionalProperties: false
      required: [vault, snapshotResealed]
      properties:
        vault: { $ref: "#/components/schemas/PasswordVaultStatus" }
        # バケットのライブのスナップショットが新しいパスワードで封じ直されたか。
        # 理由付きの false は、ローカルの変更は済んでいるが、バケットにはまだ古い
        # パスワードでしか開かないものが残っている、という意味である。
        snapshotResealed: { type: boolean }
        snapshotProblem: { type: string }
    StoreCredentialRequest:
      type: object
      additionalProperties: false
      required: [secret]
      properties:
        secret: { type: string, minLength: 1, maxLength: 1024 }
    AssignCredentialRequest:
      type: object
      additionalProperties: false
      required: [subject, name]
      properties:
        subject: { type: string, minLength: 1, maxLength: 4096 }
        name: { type: string, minLength: 1, maxLength: 128 }
...
    PasswordVaultStatus:
      type: object
      additionalProperties: false
      required: [exists, unlocked, aliases]
      properties:
        exists: { type: boolean }
        unlocked: { type: boolean }
        # パスワードが保存されているホスト。名前の集合は秘密ではない。秘密なのは
        # 値の方であり、どの操作もそれを返さない。
        aliases:
          type: array
          items: { type: string }
        minPassphraseLength: { type: integer }
    # blocker は保存済みパスワードを使用不能にするので、保存自体を拒否する。
    # warning は、ユーザーの方がよく知っているかもしれない事実である — 設定済みの
    # 鍵が向こう側で認可されていないのはよくあることだ — ので、事実として述べる
    # にとどめ、判断は本来それが属する場所に委ねる。
    PasswordEligibility:
      type: object
      additionalProperties: false
      required: [alias, storable, blockers, warnings]
      properties:
        alias: { type: string }
        storable: { type: boolean }
        blockers:
          type: array
          items: { $ref: "#/components/schemas/Notice" }
        warnings:
          type: array
          items: { $ref: "#/components/schemas/Notice" }
        hostName: { type: string }
        port: { type: string }
    PassphraseRequest:
      type: object
      additionalProperties: false
      required: [passphrase]
      properties:
        passphrase: { type: string, minLength: 1, maxLength: 1024 }
    StorePasswordRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password: { type: string, minLength: 1, maxLength: 1024 }
...
    Notice:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string }
        path: { type: string }
        line: { type: integer }
        detail: { type: string }
    Diagnostic:
      type: object
      additionalProperties: false
      required: [severity, code]
      properties:
        severity: { type: string }
        code: { type: string }
        path: { type: string }
        absolute: { type: string }
        external: { type: boolean }
        line: { type: integer }
        detail: { type: string }
...
    DiffLine:
      type: object
      additionalProperties: false
      required: [op, text]
      properties:
        op: { type: string }
        text: { type: string }
        oldLine: { type: integer }
        newLine: { type: integer }
...
    ConflictReport:
      type: object
      additionalProperties: false
      required: [path, externalChange, localChange]
      properties:
        path: { type: string }
        baseDigest: { type: string }
        diskDigest: { type: string }
        externalChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
        localChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }